requirements.txt
pywebview~=6.0
dotenv~=0.9.9
python-dotenv~=1.1.1